# Conflicts:
#	.idea/dictionaries/project.xml
#	src/main/java/ca/kpu/info2413/library/backend/DTO/PublicationDTO.java
#	src/main/java/ca/kpu/info2413/library/backend/config/SecurityConfig.java
#	src/main/java/ca/kpu/info2413/library/backend/config/WebConfig.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/AccountController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/AuthController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/BookCopyConditionNotesController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/BookCopyController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/BorrowController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/FineController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/HoldController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/PaymentController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/PublicationController.java
#	src/main/java/ca/kpu/info2413/library/backend/controller/RegistrationController.java
#	src/main/java/ca/kpu/info2413/library/backend/model/AccountPublicationId.java
#	src/main/java/ca/kpu/info2413/library/backend/model/Author.java
#	src/main/java/ca/kpu/info2413/library/backend/model/BookCopy.java
#	src/main/java/ca/kpu/info2413/library/backend/model/BookCopyConditionNotes.java
#	src/main/java/ca/kpu/info2413/library/backend/model/Borrow.java
#	src/main/java/ca/kpu/info2413/library/backend/model/LibraryCard.java
#	src/main/java/ca/kpu/info2413/library/backend/model/Publication.java
#	src/main/java/ca/kpu/info2413/library/backend/model/PublicationAuthorId.java
#	src/main/java/ca/kpu/info2413/library/backend/model/PublicationPublicationId.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/AccountPublicationRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/BookCopyRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/BorrowRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/HoldRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/LibraryCardRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/repository/PublicationRepository.java
#	src/main/java/ca/kpu/info2413/library/backend/security/AccountUserDetails.java
#	src/main/java/ca/kpu/info2413/library/backend/security/AccountUserDetailsService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/AccountService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/BookCopyService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/BorrowService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/FineService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/HoldService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/LibraryCardService.java
#	src/main/java/ca/kpu/info2413/library/backend/service/PublicationService.java
#	src/main/resources/application.properties
#	src/main/resources/static/AccountLogin.html
#	src/main/resources/static/AddDeleteUser.html
#	src/main/resources/static/AdminPage.html
#	src/main/resources/static/BorrowPage.html
#	src/main/resources/static/CatalogPage.html
#	src/main/resources/static/EditUserInformation.html
#	src/main/resources/static/HomePage.html
#	src/main/resources/static/Inventory page.html
#	src/main/resources/static/LibraryReportsDashboard.html
#	src/main/resources/static/PasswordRecoveryPage.html
#	src/main/resources/static/PaymentPage.html
#	src/main/resources/static/PlaceHold.html
#	src/main/resources/static/ReturnBook.html
#	src/main/resources/static/script.js

+++ src/main/java/ca/kpu/info2413/library/backend/model/PublicationAuthorId.java
@@ -20,5 +20,5 @@ public final class PublicationAuthorId implements Serializable
     private Integer authorIdAuthor;
 
     @Column(name = "isbn_13_Publication", nullable = false)
-    private Integer isbn13Publication;
+    private Long isbn13Publication;
 }
